Examples: Adopt the structure used in hummingbird-examples
